working collision
Collision is working, it will remove bullets and asteroids when they hit. If the ship hits the asteroid it will print a message to show they collide. I had to add 2 methods to the gameObject interface and allow GameLoop to take the game pane as a parameter to delete the visuals.